+++ docs/releases.md
@@ -5,3 +5,7 @@
    - adds to system PATH on pip install
    - supports loading csvs of questions and answers
    - basic browser of the loaded flashcards
+- 0.03: 
+   - add 10 default questions and answers
+   - implement textwrap and dynamic text centering for flashcards
+   - add top-left text to keep track of current flashcard number